requirements.txt
Flask
openai
python-dotenv
markdown2
pymongo
apscheduler
